GLibTools.cmake: Use basename in glib-mkenums templates
This change helps to create reproducible builds. Even the glib-mkenums
is not called in the sources, it'll be better to sync the GLibTools.cmake
content with the evolution-data-server, in case anybody uses this project
as a template for another module and will use the glib-mkenums.

Related to https://gitlab.gnome.org/GNOME/evolution-data-server/-/issues/281