CMake: Add FORCE_INSTALL_PREFIX compile time option
This allows to install evolution-data-server and evolution files
files into the CMAKE_INSTALL_PREFIX instead of the prefix advertised
by respective pkg-config files. It's useful, when the eds/evo prefixes
are not writable for the builder.

Also use EBookMetaBackendClass::backend_module_directory, which is
closely related to this new option.